refactor: move logo in navigation component folder
